defconfig: Enable MediaTek Scheduler Monitor Support
This commit enables the CONFIG_MTK_SCHED_MONITOR configuration option in the kernel. The option is specific to the MediaTek platform and likely introduces support for monitoring and managing the scheduler behavior on MediaTek-based devices. Enabling this feature enhances the kernel's capabilities in optimizing task scheduling and resource utilization for improved system performance.

Signed-off-by: officialputuid <officialputuid@hack.id>